Made the bundle script better

+++ bundle.sh
@@ -1,9 +1,20 @@
 #!/bin/bash
 
-rollup --config bundle-config.js
-head -n -7 bundle.js > /c/Users/lordg/AppData/Local/Screeps/scripts/screeps.com/custom/main.js
-rm bundle.js
-echo "" >> /c/Users/lordg/AppData/Local/Screeps/scripts/screeps.com/custom/main.js
-echo "module.exports.loop = loop;" >> /c/Users/lordg/AppData/Local/Screeps/scripts/screeps.com/custom/main.js
+GAME_LOCATION="$1"
+if [[ "$1" == "" ]]; then
+    GAME_LOCATION=$(cat main.js.location)
+fi
 
-touch /c/Users/lordg/AppData/Local/Screeps/scripts/screeps.com/custom/main.js
+if [[ ! -f "${GAME_LOCATION}" ]]; then
+    echo "${GAME_LOCATION} does not point to a file".
+    exit 1
+fi
+
+rollup --config bundle-config.js
+head -n -7 bundle.js > bundle.mod.js
+rm bundle.js
+echo "" >> bundle.mod.js
+echo "module.exports.loop = loop;" >> bundle.mod.js
+
+mv bundle.mod.js "${GAME_LOCATION}"
+touch "${GAME_LOCATION}" # Ensure the game picks up the changes
